pyproject.toml
[project]
name = "fastapi-meeting-service"
version = "0.1.0"
description = ""
authors = [
    {name = "wizardlee88",email = "wizardlee88@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.127.0,<0.128.0)",
    "sqlmodel (>=0.0.29,<0.0.30)",
    "sqlalchemy-utc (>=0.14.0,<0.15.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "pwdlib[argon2,bcrypt] (>=0.3.0,<0.4.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]


[tool.pytest.ini_options]
minversion = "8.3.4"
env = []
addopts = "--strict-markers --tb=short --asyncio-mode=auto -p no:warnings --doctest-modules"
python_files = ["test.py", "test_*.py"]
rootdir = "./"
testpaths = ["./tests", "./appserver"]
pythonpath = ["./"]
filterwarnings = ["error", "ignore::DeprecationWarning:dtcd3.*:"]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

